% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aavAnalysis.R
\docType{package}
\name{aavAnalysis}
\alias{aavAnalysis}
\title{aavAnalysis: analysis toolkit for AAVengeR}
\description{
Contains useful functions to analyze and generate table and plots
for AAVengeR report pipeline.
}
\section{Main functions}{

\itemize{
\item \code{\link{getAAVengerData}}: pull data from AAVengeR
\item \code{\link{getProcessedDf}}: Preprocess dataframe for analysis.
}
}

\section{Author}{

Maintainer: Tianyu Lu (\email{tianyu.lu@pennmedicine.upenn.edu})
}

\section{Links}{

GitHub \url{https://github.com/tianyulll/aavAnalysis/tree/main}

Report issues at \url{https://github.com/tianyulll/aavAnalysis/issues}
}

