% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAAVengeRData.R
\name{getAAVengerData}
\alias{getAAVengerData}
\title{Pull AAVengeR data from input or database}
\usage{
getAAVengerData(file)
}
\arguments{
\item{file}{accepts multiple format. If in RDS/excel, will be read directly.
if in string, will try to pull from aavenger database.}
}
\value{
aavenger outputs in a dataframe.
}
\description{
Get inputs from AAVengeR pipeline outputs.
It accepts input in rds or excel. It also accepts a string for trial name in aavenger database.
To use this function, the user must set up database credentials in ~/.my.cnf
}
\examples{
aavAnalysis::getAAVengerData("AAVHelaTopo")
getAAVengerData("/path/to/my/aavenger.rds")

}
