% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundanceDf.R
\name{getAbundanceDf}
\alias{getAbundanceDf}
\title{Summarize Clonal Abundance}
\usage{
getAbundanceDf(df, topn = 10)
}
\arguments{
\item{df}{df after \code{\link{getProcessedDf}}}

\item{topn}{how many top largest clone should be return. Default at 10.
Choosing values other than 10 will break the plotting function.
Will be fixed in future version.}
}
\value{
dataframe
}
\description{
Summarize Clonal Abundance per sample. Finds the Top n largest clones.
}
