% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAAVengeRData.R
\name{getAvailAAVengeR}
\alias{getAvailAAVengeR}
\title{get available trials in aavenger database}
\usage{
getAvailAAVengeR()
}
\value{
NULL. prints available trial names
}
\description{
To use this function, the user must set up database credentials in ~/.my.cnf
}
\examples{
getAvailAAVengeR()
}
