% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneDf.R
\name{getGeneDf}
\alias{getGeneDf}
\title{In gene and in TU Summary}
\usage{
getGeneDf(df)
}
\arguments{
\item{df}{df after \code{\link{getProcessedDf}}}
}
\value{
dataframe with summarized info about in_gene and in_transcription_unit columns.
}
\description{
In gene and in TU Summary
}
