% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryDf.R
\name{getProcessedDf}
\alias{getProcessedDf}
\title{summarize AAVengeR outputs}
\usage{
getProcessedDf(df, meta = NULL, minreads = 0)
}
\arguments{
\item{df}{dataframe from aavenger outputs}

\item{meta}{path to meta data in rds. If not provided, will use subject column in df}

\item{minreads}{exclude sites with reads small than minreads}
}
\value{
summarized dataframe of integration sites per sample
}
\description{
summarize AAVengeR outputs
}
