% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRearrangement.R
\name{getRearrangeDf}
\alias{getRearrangeDf}
\title{get rearrangement summary table}
\usage{
getRearrangeDf(df)
}
\arguments{
\item{df}{df after \code{\link{getProcessedDf}}}
}
\value{
dataframe with ITR rearrangement summary
}
\description{
get rearrangement summary table
}
