% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryDf.R
\name{getSummary}
\alias{getSummary}
\title{Summarize aavenger result by sample}
\usage{
getSummary(df)
}
\arguments{
\item{df}{preprocessed dataframe}
}
\value{
dataframe summarized by sample
}
\description{
Summarize aavenger result by sample
}
