% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundance.R
\name{plotListAbundance}
\alias{plotListAbundance}
\title{abundance plot for all samples}
\usage{
plotListAbundance(df, saveimg = F, outputDir = NULL)
}
\arguments{
\item{df}{dataframe from \code{\link{getAbundanceDf}}}

\item{saveimg}{boolean. If true will create directory and save the image as PNG.}

\item{outputDir}{Output Directory if saving the image.}
}
\value{
List of ggplot2 objects
}
\description{
Iterate through the abundnace dataframe by samples.
Creates abundance plot per sample
}
