% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRemnant.R
\name{plotListRemnant}
\alias{plotListRemnant}
\title{Remnant Plot for all Samples}
\usage{
plotListRemnant(
  df,
  saveimg = F,
  outDir = NULL,
  buildAAVremnantPlots_ITRseqStart = 57,
  buildAAVremnantPlots_ITRlength = 197,
  buildAAVremnantPlots_NTbinSize = 3
)
}
\arguments{
\item{df}{dataframe from \code{\link{getProcessedDf}}}

\item{saveimg}{Default false. whether saving the image as PNG.}

\item{outDir}{path to save img.}

\item{buildAAVremnantPlots_ITRseqStart}{plot parameter: ITR start position. Default 57}

\item{buildAAVremnantPlots_ITRlength}{plot parameter: ITR length. Default 197}

\item{buildAAVremnantPlots_NTbinSize}{plot parameter: ITR bin size. Default 3}
}
\value{
List of ggplot2 objects.
}
\description{
Remnant Plot for all Samples
}
