% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotRandomGene}
\alias{plotRandomGene}
\title{Plot Random Site in Gene and in TU}
\usage{
plotRandomGene(df, random_exon = NULL, random_tu = NULL)
}
\arguments{
\item{df}{df after \code{\link{getGeneDf}}}

\item{random_exon}{random sites in Exon percentange}

\item{random_tu}{random sites in Transcription Unit Percentage}
}
\value{
ggplot2 object
}
\description{
Plot Random Site in Gene and in TU
}
