% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRearrangement.R
\name{plotRearrangment}
\alias{plotRearrangment}
\title{Plot Rearrangment Percentage Summary}
\usage{
plotRearrangment(df)
}
\arguments{
\item{df}{output df from  \code{\link{getRearrangeDf}}}
}
\value{
ggplot2 object
}
\description{
Plot Rearrangment Percentage Summary
}
