% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundance.R
\name{plotSampleAbundance}
\alias{plotSampleAbundance}
\title{Plot Abundance of one sample}
\usage{
plotSampleAbundance(df, saveimg = F, outputDir = NULL)
}
\arguments{
\item{df}{One sample(GTSP) from the dataframe after \code{\link{getAbundanceDf}}}

\item{saveimg}{boolean. If true will create directory and save the image as PNG.}

\item{outputDir}{Output Directory if saving the image.}
}
\value{
ggplot2 object of the abundance plot
}
\description{
First, Calculate the rest/unselected clonotypes as low abundance.
Second, Sort the clones by its size.
Finally, generate clonal abundance plots and save if flagged.
}
