% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRemnant.R
\name{plotSampleRemnant}
\alias{plotSampleRemnant}
\title{Remnant plot for one sample}
\usage{
plotSampleRemnant(
  x,
  saveimg = F,
  outDir = NULL,
  buildAAVremnantPlots_ITRseqStart = 57,
  buildAAVremnantPlots_ITRlength = 197,
  buildAAVremnantPlots_NTbinSize = 3
)
}
\arguments{
\item{x}{dataframe per sample}

\item{saveimg}{Default false. whether saving the image as PNG.}

\item{outDir}{path to save img.}

\item{buildAAVremnantPlots_ITRseqStart}{plot parameter: ITR start position. Default 57}

\item{buildAAVremnantPlots_ITRlength}{plot parameter: ITR length. Default 197}

\item{buildAAVremnantPlots_NTbinSize}{plot parameter: ITR bin size. Default 3}
}
\value{
ggplot2 object
}
\description{
Remnant plot for one sample
}
