% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDBConfig.R
\name{setDbConfig}
\alias{setDbConfig}
\title{Writes database connection}
\usage{
setDbConfig(
  user = "aavenger",
  password,
  host = "174.129.238.44",
  port = "3306",
  database = "AAVengeR",
  file_path = "~/.my.cnf"
)
}
\arguments{
\item{password}{to access the database}
}
\description{
By default, writes aavenger read credentials to ~/.my.cnf
must give password.
}
