% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotRandomGene}
\alias{plotRandomGene}
\title{Plot Random Site in Gene and in TU}
\usage{
plotRandomGene(df, exon_mean = 0, exon_sd = 0, tu_mean = 0, tu_sd = 0)
}
\arguments{
\item{df}{df after \code{\link{getGeneDf}}}

\item{exon_mean}{mean \% of random site in exons}

\item{exon_sd}{std \% of random site in exons}

\item{tu_mean}{mean \% of random site in transcription unit}

\item{tu_sd}{std \% of random site in transcription unit}
}
\value{
ggplot2 object
}
\description{
Plot Random Site in Gene and in TU
}
